+++ lib/pages/global-search-screen.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class GlobalSearchScreen extends StatefulWidget {
+  _GlobalSearchScreenState createState() => _GlobalSearchScreenState();
+}
+
+class _GlobalSearchScreenState extends State<GlobalSearchScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Container(
+        child: Text('Hello World'),
+      ),
+    );
+  }
+}