+++ lib/pages/others/help-screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
 
-import 'package:moblie/widgets/ui/list_cell.dart';
-import 'package:moblie/widgets/ui/text/subtitle.dart';
-import 'package:moblie/widgets/ui/text/title.dart';
+import 'package:vocabulary_builder/widgets/ui/list_cell.dart';
+import 'package:vocabulary_builder/widgets/ui/text/subtitle.dart';
+import 'package:vocabulary_builder/widgets/ui/text/title.dart';
 
 class HelpScreen extends StatelessWidget {