+++ lib/pages/screens/intro-screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
-
-import 'package:moblie/model/main.dart';
 import 'package:scoped_model/scoped_model.dart';
-import 'package:moblie/widgets/ui/elements/slider.dart';
+
+import 'package:vocabulary_builder/model/main.dart';
+import 'package:vocabulary_builder/widgets/ui/elements/slider.dart';
 
 class IntroScreen extends StatelessWidget {