+++ lib/pages/screens/welcome-screen.dart
@@ -1,94 +1,98 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
-import 'package:moblie/model/main.dart';
-import 'package:moblie/utils/colors.dart';
-import 'package:moblie/widgets/ui/divider.dart';
+import 'package:vocabulary_builder/model/main.dart';
+
 import 'package:scoped_model/scoped_model.dart';
 
-class WelcomeScreen extends StatelessWidget {
+class WelcomeScreen extends StatefulWidget {
+  @override
+    State<StatefulWidget> createState() => _WelcomeScreenState();
+}
+
+class _WelcomeScreenState extends State<WelcomeScreen> {
   @override
   Widget build(BuildContext context) {
     return ScopedModelDescendant<MainModel> (
       builder: (BuildContext context, Widget child, MainModel model) {
         return Scaffold(
-          body: Container(
-
-            width: double.infinity,
-            height: double.infinity,
-            child: Column(
-              children: <Widget>[
-              
-                _renderText(
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 30.0,
-                  ),
-
-                  title: FlutterI18n.translate(context, 'welcome_screen.title'),
-                  margin: EdgeInsets.only(top: 150.0, bottom: 10.0, left:     20, right: 20),
+          body: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Container(
+                child: Stack(
+                  children: <Widget>[
+                    Container(
+                      padding: EdgeInsets.only(top: 120, left: 20.0, right: 20.0),
+                      child: Text(FlutterI18n.translate(context, 'welcome_screen.title'),
+                        style: TextStyle(
+                          fontSize: 50.0, 
+                          fontWeight: FontWeight.bold
+                        )
+                      ),
+                    ),
+                  ],
                 ),
-
-                _renderText(
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 18.0,
-                  ),
-
-                  title: FlutterI18n.translate(context, 'welcome_screen.subtitle'),
-                  margin: EdgeInsets.only(top: 20.0, bottom: 30.0, left: 20,     right: 20)
+              ),
+            
+              Container(
+                padding: EdgeInsets.only(top: 80.0, left: 20.0, right: 20.0),
+                child: Column(
+                  children: <Widget>[
+                    _renderButton(
+                      name: 'Spanish',
+                      language: 'en',
+                      model: model,
+                      context: context
+                    ),
+                    SizedBox(height: 20.0),
+                    _renderButton(
+                      name: 'Inglés',
+                      language: 'es',
+                      model: model,
+                      context: context
+                    ),
+                  ],
                 ),
-
-                Separator.spacer(height: 20.0),
-                _renderButton(
-                  name: 'Inglés',
-                  language: 'es',
-                  model: model,
-                  context: context
-                ),
-
-                Separator.spacer(height: 20.0),
-                _renderButton(
-                  name: 'Español',
-                  language: 'en',
-                  model: model,
-                  context: context
-                )
-              ],
-            ),
-          )
+              ),
+            ],
+          ),
         );
       },
     );
   }
 
   Widget _renderButton({String name, String language, model, context}) {
-    return RaisedButton(
-      color: transparent,
-      child: Container(
-        alignment: Alignment.center,
-        child: Text(name),
-        width: 200,
-        height: 60,
-      ),
-      onPressed: () {
-        model.setUserPreferedLanguage(language: language);
+    return GestureDetector(
 
-        Navigator.of(context).pushReplacementNamed('/intro');
+      onTap: () async {
+        
+        setState(() => model.setUserPreferedLanguage(language: language));
+        
+        await model.obtenerData();
+        await model.seen
+          ? Navigator.pushReplacementNamed(context, '/home') 
+          : Navigator.pushReplacementNamed(context, '/intro');
       },
-    );
-  }
 
-  Widget _renderText({String title, TextStyle style, EdgeInsetsGeometry margin}) {
-    return Container(
-      child: Text(
-        title,
-        style: style,
-        textAlign: TextAlign.center,
+      child: Container(
+        height: 52.0,
+        child: Material(
+          borderRadius: BorderRadius.circular(20.0),
+          color: Color(0xFFFFBC02D),
+          elevation: 7.0,
+          child:Center(
+            child: Text(
+              name,
+              style: TextStyle(
+                fontSize:20.0, 
+                color: Colors.white,
+                fontWeight: FontWeight.bold,
+                fontFamily: 'Poppins'
+              ),
+            ),
+          ),
+        )
       ),
-
-      margin: margin,
     );
   }
-
 }