+++ lib/widgets/ui/head-card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
-import 'package:moblie/model/main.dart';
-import 'package:moblie/utils/colors.dart';
-import 'package:moblie/widgets/ui/divider.dart';
+import 'package:vocabulary_builder/model/main.dart';
+import 'package:vocabulary_builder/utils/colors.dart';
+import 'package:vocabulary_builder/widgets/ui/divider.dart';
 import 'package:clipboard_manager/clipboard_manager.dart';
 
 class HeadCard extends StatelessWidget {